requirements.txt
requests = "*"
beautifulsoup4 = "*"
llama-index = "0.10.0"
llama-index-core = "*"
openai = "*"
python-dotenv = "*"
pinecone-client = "*"
black = "*"
streamlit = "*"
watchdog = "*"
setuptools = "*"